Fix docs static build issues
