+++ ui/homepage/app/docs/[collection]/[version]/page.tsx
@@ -4,7 +4,7 @@ import { notFound } from 'next/navigation'
 import type { Metadata } from 'next'
 
 import Breadcrumbs, { type Crumb } from '../../../../components/download/Breadcrumbs'
-import { DOC_COLLECTIONS, getDocResource } from '../../resources'
+import { DOC_COLLECTIONS, getDocResource } from '../../resources.server'
 import { isFeatureEnabled } from '@lib/featureToggles'
 import DocCollectionView from './DocCollectionView'
