+++ ui/homepage/app/docs/resources.server.ts
@@ -4,76 +4,8 @@ import { isFeatureEnabled } from '@lib/featureToggles'
 import docsManifest from '../../public/dl-index/docs-manifest.json'
 import fallbackDocsIndex from '../../public/_build/docs_index.json'
 
-const DEFAULT_DOCS_BASE_URL = 'https://dl.svc.plus/docs'
-
-const normalizeBaseUrl = (value?: string) => {
-  if (!value) return ''
-  const trimmed = value.trim()
-  if (!trimmed) return ''
-  return trimmed.replace(/\/$/, '')
-}
-
-const docsBaseUrl = normalizeBaseUrl(process.env.NEXT_PUBLIC_DOCS_BASE_URL) || DEFAULT_DOCS_BASE_URL
-
-export const DOCS_BASE_URL = docsBaseUrl
-
-export const buildAbsoluteDocUrl = (value?: string) => {
-  if (!value || typeof value !== 'string') {
-    return undefined
-  }
-  const trimmed = value.trim()
-  if (!trimmed) {
-    return undefined
-  }
-  if (/^https?:\/\//i.test(trimmed)) {
-    return trimmed
-  }
-
-  if (!docsBaseUrl) {
-    return trimmed.startsWith('/') ? trimmed : `/${trimmed}`
-  }
-
-  try {
-    const base = new URL(docsBaseUrl.endsWith('/') ? docsBaseUrl : `${docsBaseUrl}/`)
-    const basePath = base.pathname.replace(/\/+$/, '')
-    const basePathWithoutLeadingSlash = basePath.replace(/^\/+/, '')
-
-    let relative = trimmed.replace(/^\/+/, '')
-
-    if (
-      basePathWithoutLeadingSlash &&
-      relative.toLowerCase().startsWith(`${basePathWithoutLeadingSlash.toLowerCase()}/`)
-    ) {
-      relative = relative.slice(basePathWithoutLeadingSlash.length + 1)
-    }
-
-    return new URL(relative || '.', base).toString()
-  } catch {
-    const ensureLeadingSlash = trimmed.startsWith('/') ? trimmed : `/${trimmed}`
-    return `${docsBaseUrl.replace(/\/+$/, '')}${ensureLeadingSlash}`
-  }
-}
-
-export interface DocResource {
-  slug: string
-  title: string
-  description: string
-  category?: string
-  version?: string
-  updatedAt?: string
-  pdfUrl?: string
-  htmlUrl?: string
-  tags?: string[]
-  estimatedMinutes?: number
-  coverImage?: string
-  language?: string
-  variant?: string
-  versionSlug?: string
-  pathSegments?: string[]
-  collection?: string
-  collectionSlug?: string
-  collectionLabel?: string
-}
+import { buildAbsoluteDocUrl } from './utils'
+import type { DocCollection, DocResource, DocVersionOption } from './types'
 
 interface RawDocResource {
   slug?: unknown
@@ -105,29 +37,6 @@ export const DOCS_DATASET = RAW_DOCS.map((item) => normalizeResource(item as Raw
   (item): item is DocResource => item !== null,
 )
 
-export interface DocVersionOption {
-  id: string
-  label: string
-  resource: DocResource
-  slug: string
-  pathSegment?: string
-}
-
-export interface DocCollection {
-  slug: string
-  title: string
-  description: string
-  category?: string
-  updatedAt?: string
-  estimatedMinutes?: number
-  tags: string[]
-  latestVersionLabel?: string
-  latestVariant?: string
-  versions: DocVersionOption[]
-  defaultVersionId?: string
-  defaultVersionSlug?: string
-  directory?: string
-}
 
 function slugifySegment(value: string): string {
   const base = value
